Accomplished - Getting a soldier with click
All the buttons in the buttons board became a new type of buttons - "SoldierButton" - its our type of button (inherits from Button) and have a property called "Position" which represents its position in the matrix.

each click on a button on the buttons board gets the soldier that sits in that position on the soldiers board, and null if empty.