Readability improvments + game ending + score board
Found 1 engine bug - if player tried to commit an eat move but he doesn't have any, the game crashes. added a condition to the sortMoveType method, need to inspect again though.

changed the text of buttons to X and O - the old code is still in comments

Divided button_Click into different methods

Each player label now holds the name and the score of the player

added the game ending method - the idea is the same as before: we show who won, ask for another game, and if the user wants another game we go toi initializeAnotherGame() method. If the user wants to quit, we show a good bye message and close the form.

initializeANotherGame - here we dispose the game buttons board and the labels, we re set the logic part (as we did in ex.02 as wel) and we call initializeButtonsBoard() again to set the new game form

in addition, changes some variables names and methods to improve code readability

luv luv luv